Algunas tablas de tiempo. Faltan definir algunas cosas para poder continuar

+++ 20161Q_GR04/SCRIPT_CREACION_TABLAS.sql
@@ -1,8 +1,18 @@
 USE [MIS_GRUPO_04]
 GO
 
-DROP TABLE dbo.LK_Estacion
-DROP TABLE dbo.LK_Linea
+IF OBJECT_ID('dbo.LK_Mes', 'U') IS NOT NULL 
+  DROP TABLE dbo.LK_Mes; 
+
+IF OBJECT_ID('dbo.LK_Anio', 'U') IS NOT NULL 
+  DROP TABLE dbo.LK_Anio; 
+
+IF OBJECT_ID('dbo.LK_Estacion', 'U') IS NOT NULL 
+  DROP TABLE dbo.LK_Estacion; 
+  
+IF OBJECT_ID('dbo.LK_Linea', 'U') IS NOT NULL 
+  DROP TABLE dbo.LK_Linea; 
+
 
 CREATE TABLE [dbo].[LK_Linea](
 	[Linea_SK] [int] IDENTITY(1,1) NOT NULL,
@@ -38,3 +48,61 @@ GO
 
 ALTER TABLE [dbo].[LK_Estacion] CHECK CONSTRAINT [FK_LK_Estacion_LK_Linea]
 GO
+
+CREATE TABLE [dbo].[LK_Anio](
+	[Anio_SK] [int] IDENTITY(1,1) NOT NULL,
+	[Anio_numero] [int] NOT NULL,
+ CONSTRAINT [PK_LK_Anio] PRIMARY KEY CLUSTERED 
+(
+	[Anio_SK] ASC
+)WITH (PAD_INDEX  = OFF, STATISTICS_NORECOMPUTE  = OFF, IGNORE_DUP_KEY = OFF, ALLOW_ROW_LOCKS  = ON, ALLOW_PAGE_LOCKS  = ON) ON [PRIMARY]
+) ON [PRIMARY]
+
+
+--Inserto los anios
+INSERT INTO dbo.LK_Anio (Anio_numero) VALUES(2013),(2014),(2015),(2016),(2017)
+
+CREATE TABLE [dbo].[LK_Mes](
+	[Mes_SK] [int] IDENTITY(1,1) NOT NULL,
+	[Mes_Numero] [int] NOT NULL,
+	[Anio_SK] [int] NOT NULL,
+ CONSTRAINT [PK_LK_Mes] PRIMARY KEY CLUSTERED 
+(
+	[Mes_SK] ASC
+)WITH (PAD_INDEX  = OFF, STATISTICS_NORECOMPUTE  = OFF, IGNORE_DUP_KEY = OFF, ALLOW_ROW_LOCKS  = ON, ALLOW_PAGE_LOCKS  = ON) ON [PRIMARY]
+) ON [PRIMARY]
+
+GO
+
+ALTER TABLE [dbo].[LK_Mes]  WITH CHECK ADD  CONSTRAINT [FK_LK_Mes_LK_Anio] FOREIGN KEY([Anio_SK])
+REFERENCES [dbo].[LK_Anio] ([Anio_SK])
+GO
+
+ALTER TABLE [dbo].[LK_Mes] CHECK CONSTRAINT [FK_LK_Mes_LK_Anio]
+GO
+
+--Inserto los meses de cada anio
+INSERT INTO dbo.LK_Mes (Mes_Numero, Anio_SK)
+	SELECT 1, Anio_SK FROM dbo.LK_Anio
+	UNION ALL
+	SELECT 2, Anio_SK FROM dbo.LK_Anio
+	UNION ALL
+	SELECT 3, Anio_SK FROM dbo.LK_Anio
+	UNION ALL
+	SELECT 4, Anio_SK FROM dbo.LK_Anio
+	UNION ALL
+	SELECT 5, Anio_SK FROM dbo.LK_Anio
+	UNION ALL
+	SELECT 6, Anio_SK FROM dbo.LK_Anio
+	UNION ALL
+	SELECT 7, Anio_SK FROM dbo.LK_Anio
+	UNION ALL
+	SELECT 8, Anio_SK FROM dbo.LK_Anio
+	UNION ALL
+	SELECT 9, Anio_SK FROM dbo.LK_Anio
+	UNION ALL
+	SELECT 10, Anio_SK FROM dbo.LK_Anio
+	UNION ALL
+	SELECT 11, Anio_SK FROM dbo.LK_Anio
+	UNION ALL
+	SELECT 12, Anio_SK FROM dbo.LK_Anio
